chore(overview): EnumCellValue component
